# [pytorch][vulkan] Add unused function annotation in vulkan perf test